Directly use returend fieldChanges from set/Clear entry Minimize test
